static cast

+++ tasks/mpi/beresnev_a_min_values_by_matrix_columns/src/ops_mpi.cpp
@@ -20,8 +20,8 @@ bool beresnev_a_min_values_by_matrix_columns_mpi::TestMPITaskSequential::validat
   // Check count elements of output
   return taskData->inputs_count[0] > 0 &&
          taskData->inputs_count[0] == reinterpret_cast<std::vector<int>*>(taskData->inputs[0])[0].size() &&
-         taskData->inputs_count[0] == (uint32_t) reinterpret_cast<int*>(taskData->inputs[1])[0] *
-                                          reinterpret_cast<int*>(taskData->inputs[2])[0] &&
+         taskData->inputs_count[0] == static_cast<uint32_t>(reinterpret_cast<int*>(taskData->inputs[1])[0]) *
+                                          static_cast<uint32_t>(reinterpret_cast<int*>(taskData->inputs[2])[0]) &&
          taskData->outputs_count[0] == reinterpret_cast<std::vector<int>*>(taskData->outputs[0])[0].size();
 }
 
@@ -68,8 +68,8 @@ bool beresnev_a_min_values_by_matrix_columns_mpi::TestMPITaskParallel::validatio
   internal_order_test();
   if (world.rank() == 0) {
     return taskData->inputs_count[0] > 0 &&
-           taskData->inputs_count[0] == (uint32_t) reinterpret_cast<int*>(taskData->inputs[1])[0] *
-                                            reinterpret_cast<int*>(taskData->inputs[2])[0] &&
+           taskData->inputs_count[0] == static_cast<uint32_t>(reinterpret_cast<int*>(taskData->inputs[1])[0]) *
+                                            static_cast<uint32_t>(reinterpret_cast<int*>(taskData->inputs[2])[0]) &&
            taskData->outputs_count[0] == (uint32_t) reinterpret_cast<int*>(taskData->inputs[2])[0] &&
            taskData->inputs_count[1] == 1 && taskData->inputs_count[2] == 1;
   }
